合并dev分支至master

See merge request cosmo/mva-cloud!103

+++ cloud-service/cloud-system/src/main/java/com/cloud/system/controller/SysOssController.java
@@ -5,25 +5,26 @@ import com.cloud.common.auth.annotation.HasPermissions;
 import com.cloud.common.core.controller.BaseController;
 import com.cloud.common.core.domain.R;
 import com.cloud.common.exception.file.OssException;
+import com.cloud.common.utils.StringUtils;
 import com.cloud.common.utils.ValidatorUtils;
+import com.cloud.common.utils.file.FileUtils;
 import com.cloud.system.domain.entity.SysOss;
-import com.cloud.system.oss.CloudConstant;
+import com.cloud.system.oss.*;
 import com.cloud.system.oss.CloudConstant.CloudService;
-import com.cloud.system.oss.CloudStorageConfig;
-import com.cloud.system.oss.CloudStorageService;
-import com.cloud.system.oss.OSSFactory;
 import com.cloud.system.oss.valdator.AliyunGroup;
-import com.cloud.system.oss.valdator.QcloudGroup;
-import com.cloud.system.oss.valdator.QiniuGroup;
+import com.cloud.system.oss.valdator.HuaweiGroup;
 import com.cloud.system.service.ISysConfigService;
 import com.cloud.system.service.ISysOssService;
 import com.google.gson.Gson;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.Date;
+import java.util.List;
 
 /**
  * 文件上传 提供者
@@ -33,6 +34,7 @@ import java.util.Date;
  */
 @RestController
 @RequestMapping("oss")
+@Api("文件上传")
 public class SysOssController extends BaseController {
     private final static String KEY = CloudConstant.CLOUD_STORAGE_CONFIG_KEY;
 
@@ -62,15 +64,12 @@ public class SysOssController extends BaseController {
     public R saveConfig(CloudStorageConfig config) {
         // 校验类型
         ValidatorUtils.validateEntity(config);
-        if (config.getType() == CloudService.QINIU.getValue()) {
-            // 校验七牛数据
-            ValidatorUtils.validateEntity(config, QiniuGroup.class);
+        if (config.getType() == CloudService.HUAWEIYUN.getValue()) {
+            // 校验华为云数据
+            ValidatorUtils.validateEntity(config, HuaweiGroup.class);
         } else if (config.getType() == CloudService.ALIYUN.getValue()) {
             // 校验阿里云数据
             ValidatorUtils.validateEntity(config, AliyunGroup.class);
-        } else if (config.getType() == CloudService.QCLOUD.getValue()) {
-            // 校验腾讯云数据
-            ValidatorUtils.validateEntity(config, QcloudGroup.class);
         }
         return toAjax(sysConfigService.updateValueByKey(KEY, new Gson().toJson(config)));
     }
@@ -103,12 +102,13 @@ public class SysOssController extends BaseController {
 
     /**
      * 修改保存文件上传
-     *
+     * @param file 文件
+     * @param orderNo 订单号
+     * @return 上传是否成功
      * @throws IOException
      */
     @PostMapping("upload")
-    @HasPermissions("sys:oss:add")
-    public R editSave(@RequestParam("file") MultipartFile file) throws IOException {
+    public R editSave(@RequestPart("file") MultipartFile file,@RequestParam(value = "orderNo",required = false) String orderNo) throws IOException {
         if (file.isEmpty()) {
             throw new OssException("上传文件不能为空");
         }
@@ -125,15 +125,84 @@ public class SysOssController extends BaseController {
         ossEntity.setFileName(fileName);
         ossEntity.setCreateTime(new Date());
         ossEntity.setService(storage.getService());
+        ossEntity.setOrderNo(orderNo);
         return toAjax(sysOssService.insertSysOss(ossEntity));
     }
 
     /**
+     * 下载文件
+     *
+     */
+    @PostMapping("downLoad")
+    public void downLoad(String url,String fileName) throws IOException {
+        String realName = new String();
+        if(StringUtils.isNotBlank(fileName)){
+            realName=fileName;
+        }else{
+            realName = url;
+        }
+        CloudStorageService storage = OSSFactory.build();
+
+        getResponse().setCharacterEncoding("utf-8");
+        // 下载使用"application/octet-stream"更标准
+        getResponse().setContentType("application/octet-stream");
+        getResponse().setHeader("Content-Disposition",
+                "attachment;filename=" + FileUtils.setFileDownloadHeader(getRequest(), realName));
+        storage.downLoad(url,getResponse().getOutputStream());
+    }
+
+    /**
      * 删除文件上传
+     * @param ids 需要删除的数据ID
      */
     @PostMapping("remove")
-    @HasPermissions("sys:oss:remove")
     public R remove(String ids) {
+        if(ids==null){
+            return R.error("参数为空!!");
+        }
+        CloudStorageService storage = OSSFactory.build();
+        for(String id:ids.split(",")){
+            SysOss sysOss = sysOssService.selectSysOssById(Long.parseLong(id));
+            storage.deleteFile(sysOss.getUrl());
+        }
         return toAjax(sysOssService.deleteSysOssByIds(ids));
     }
+
+    /**
+     * 根据订单编号查询文件上传列表
+     * @param orderNo 订单编号
+     * @return List<SysOss> 文件上传集合
+     */
+    @GetMapping("listByOrderNo")
+    @ApiOperation(value = "根据订单编号查询文件上传列表",response = SysOss.class)
+    public R listByOrderNo(@RequestParam("orderNo") String orderNo) {
+        List<SysOss> listResult = sysOssService.selectSysOssListByOrderNo(orderNo);
+        return R.data(listResult);
+    }
+
+    /**
+     * 根据订单编号修改文件上传列表
+     * @param orderNo 订单编号
+     * @param files 文件数组
+     * @return 成功或失败
+     */
+    @PostMapping("updateListByOrderNo")
+    @ApiOperation(value = "根据订单编号修改文件上传列表",response = SysOss.class)
+    public R updateListByOrderNo(@RequestParam("orderNo") String orderNo,@RequestPart("files") MultipartFile[] files) throws IOException {
+        R result = sysOssService.updateListByOrderNo(orderNo,files);
+        return result;
+    }
+
+    /**
+     * 根据订单编号删除文件上传列表
+     * @param orderNo 订单编号
+     * @return 成功或失败
+     */
+    @PostMapping("deleteListByOrderNo")
+    @ApiOperation(value = "根据订单编号删除文件上传列表",response = SysOss.class)
+    public R deleteListByOrderNo(@RequestParam("orderNo") String orderNo) {
+        R result = sysOssService.deleteListByOrderNo(orderNo);
+        return result;
+    }
+
 }
